♻️ 👽️Change the entrypoint path in docker compose
After the changes to roll forward to the recommended image,
fa2fc02e355f946480516ed31044b7922b23dc45
the build directory is in `/ext/build` and not
`/ext/source/build`.

In this commit, we change all the entrypoint scripts to point
to the new location. This is a requirement for the demos to work after fa2fc02e355f946480516ed31044b7922b23dc45

Signed-off-by: Shankari <k.shankari@ee.doe.gov>